chore: improve slap by adding hand picture

+++ dbmodel.sql
@@ -43,7 +43,7 @@ CREATE TABLE IF NOT EXISTS `card` (
   `card_type` varchar(16) NOT NULL,
   `card_type_arg` int(11) NOT NULL,
   `card_location` varchar(16) NOT NULL,
-  `card_location_arg` int(11) NOT NULL,
+  `card_location_arg` BIGINT NOT NULL,
   `hidden` BOOLEAN NOT NULL DEFAULT 0,
   PRIMARY KEY (`card_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 AUTO_INCREMENT=1 ;
